[2026-06-18] 권한 분리, mapper id, departmentId 데이터타입 long 변경

+++ src/main/java/com/msa4lms/domain/auth/mapper/AuthMapper.java
@@ -6,7 +6,7 @@ import org.apache.ibatis.annotations.Param;
 @Mapper
 public interface AuthMapper {
     int updateRefreshToken(
-            @Param("id") int id,
+            @Param("id") long id,
             @Param("refreshToken") String refreshToken
     );
 }
